Add files via upload
Bot was redesigned and now uses fewer lines of code. At the moment it reaches the destination search page but is not programmed to like pictures as of now. Unlike the original, this one can actually type in the search box and search for the term (hashtag); simulating a human searching for a tag. The functionality to like pictures is currently in the works.